FEATURE: easily extract arguments in jobs
Because of the serialization method we use (JSON), we can lose type information when we create jobs. For instance, if we enqueue a job with integer arguments, they're serialized as JSON with a Number type. When JSON decodes it, it often gets deserialized as a float64. To address this, the job itself will need to re-assert which type it expects. The expected usage of this is as follows:

func myJobHandler(job *work.Job) error {
    subdomainId := job.ArgInt64(subdomain_id)
    adminId := job.ArgInt64(admin_id)
    sendEmail := job.ArgBool(send_email)
    if err := job.ArgError(); err != nil {
        return err
    }

    // …
}

Note that the error is only checked once at the end -- this design assumes the jobs are inserted by static Go code, not user input. It also doesn't optimize for optional arguments, and assumes jobs have fixed inputs (like Go functions).