FEATURE: delete scheduled/retry jobs
This lets you, for instance: schedule a job to email a user in 5 minutes, while providing the user the option to 'undo' the send of that email. If they click undo, just delete the scheduled job, and nothing will be sent. This takes into account unique jobs as well -- deleting a scheduled unique job will clear the uniqueness for the job name/args.